Patron sudo modülünü iyileştir
