+++ app/controllers/application_controller.rb
@@ -2,7 +2,7 @@
 
 class ApplicationController < ActionController::Base
   include Pundit
-  include Patron::Sudo
+  include Patron::Sudoable
 
   protect_from_forgery with: :exception
