+++ app/lib/patron.rb
@@ -7,6 +7,7 @@ require_relative 'patron/scope'
 require_relative 'patron/permission_builder'
 require_relative 'patron/role_builder'
 require_relative 'patron/sudo'
+require_relative 'patron/sudoable'
 
 module Patron
   class Error < StandardError; end
